+++ backend/server.js
@@ -2,6 +2,7 @@ if (process.env.NODE_ENV !== 'production') { // can load .env file while server 
     require('dotenv').config();
 }
 
+const authHelper = require('./authHelper')
 const prisma = require('./database/db')
 const express = require('express') 
 const app = express()
@@ -31,14 +32,16 @@ app.get('/', (request, response) => {
     response.send ('my roots?');
 });
 
-app.post('/sign-up', async (request,response) => { // Sign up route, get the new user data 
+app.post('/sign-up', async (request,response) => { 
     const { email, password, confPass } = request.body;
 
-    if(!email || !password) return response.status(400).send('username or password are missing') // do something prettier w UI
+    if(!email || !password) {
+        return response.status(400).send('username or password are missing') // do something prettier w UI
+    }
 
     const currentUser = await prisma.user.findFirst({
         where: {
-            email: email
+            email
         },
     })
 
@@ -47,23 +50,11 @@ app.post('/sign-up', async (request,response) => { // Sign up route, get the new
     }
 
     if(confPass!=password) return response.status(401).send('passwords are not matching');
-
-    function isPasswordSecure(password) {
-        const minLength = 8;  
-        const hasUpperCase = /[A-Z]/.test(password);  
-        const hasLowerCase = /[a-z]/.test(password); 
-        const hasDigit = /\d/.test(password);  
-        const hasSpecialChar = /[!@#$%^&*(),.?":{}|<>]/.test(password); 
-
-        return (
-            password.length >= minLength && hasUpperCase && hasLowerCase && hasDigit && hasSpecialChar
-        );
-    }
     
-    if(!isPasswordSecure(password)){
+    if(!authHelper.isPasswordSecure(password)){
         return response.status(400).send('password is not secure')
     }
-    
+
     const salt = await bcrypt.genSalt(10)
     const hashedPassword = await bcrypt.hash(password, salt)
 
@@ -75,17 +66,17 @@ app.post('/sign-up', async (request,response) => { // Sign up route, get the new
         }
     })
     /* TODO */
-    response.redirect('/')
+    return response.status(200).send('sign up successful') 
 })
 
-app.post('/sign-in', async (request, response) => { // Sign in route, get the users data to login
+app.post('/sign-in', async (request, response) => { 
     const { email, password } = request.body;
 
     if(!email || !password) return response.status(400).send('Username and password are missing');
 
     const currentUser = await prisma.user.findFirst({
         where: {
-            email: email,
+            email
         }
     })
 
@@ -94,7 +85,7 @@ app.post('/sign-in', async (request, response) => { // Sign in route, get the us
 
        if(passwordMatch){
             /* TODO */
-            response.redirect('/') 
+            return response.status(200).send('login successful') 
        }else{
             return response.status(400).send('invalid email or password')
        }
